refactor(tests): ship producer-contract suite inside the wheel
Move tests/test_producer_contracts.py to
src/eigsep_observing/contract_tests/ so the suite is reachable via
`pytest --pyargs eigsep_observing.contract_tests` on installs that
only have the wheel (the Pi). The eigsep-field CLI's `verify` command
invokes exactly this path — previously it relied on a filesystem
fallback that only worked when the eigsep_observing repo was cloned
next to eigsep-field.

Extract the golden data constants (HEADER, IMU_READING, CORR_METADATA,
VNA_METADATA, S11_HEADER, NTIMES, ERROR_INTEGRATION_INDEX and the
_*_avg_entry helpers they depend on) from tests/conftest.py into a new
private module src/eigsep_observing/_test_fixtures.py. The leading
underscore marks it as non-public API — shape may change without a
deprecation cycle. tests/conftest.py re-exports the names so existing
`from conftest import HEADER` imports in test_io.py keep working.

Add src/eigsep_observing/contract_tests to pytest testpaths so plain
`pytest` from the repo root still auto-discovers the suite.
